[Arithmetic-Coding] fix overflow

+++ Arithmetic-Coding/src/coding-arithmetic-ppm.hpp
@@ -12,9 +12,9 @@ namespace coding {
 class ArithmeticPPM: public Base {
 public:
     inline static const std::string TYPE = "arithmetic-ppm";
-    using History = std::deque<uint16_t>;
+    using History = std::deque<uint32_t>;
 
-    const uint16_t bits, charset;
+    const uint32_t bits, charset;
     const int order;
 
     Arithmetic::Accum base;
@@ -24,7 +24,7 @@ public:
     ~ArithmeticPPM();
 
     Context* find(const History&);
-    void update(const History&, uint16_t);
+    void update(const History&, uint32_t);
 
     size_t encode(DataSrc&, DataDst&) final;
     size_t decode(DataSrc&, DataDst&) final;
